% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saeHB.TF.beta-package.R
\docType{package}
\name{saeHB.TF.beta-package}
\alias{saeHB.TF.beta-package}
\alias{saeHB.TF.beta}
\title{The 'saeHB.TF.beta' package.}
\description{
A DESCRIPTION OF THE PACKAGE
}
\references{
Stan Development Team (NA). RStan: the R interface to Stan. R package version 2.36.0.9000. https://mc-stan.org
}
\author{
\strong{Maintainer}: Nasya Zahira Putri \email{nasyazp28@gmail.com}

}
