% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saeHB.TF.beta-package.R
\name{saeHB.TF.beta-package}
\alias{saeHB.TF.beta-package}
\alias{saeHB.TF.beta}
\title{The 'saeHB.TF.beta' package.}
\description{
Small Area Estimation using Hierarchical Bayes Twofold Subarea Level Model under Beta Distribution
}
\references{
Stan Development Team (NA). RStan: the R interface to Stan. R package version 2.36.0.9000. https://mc-stan.org
}
